add tracking user location function && refactor model for timezone parameter

+++ data/src/main/java/com/e/data/entity/chosen/ChosenCityTemp.kt
@@ -18,9 +18,9 @@ data class ChosenCityTemp(
     val tempMax: Float,
 
     @Json(name = "pressure")
-    val pressure: Float,
+    val pressure: Int,
 
     @Json(name = "humidity")
-    val humidity: Float,
+    val humidity: Int,
 )
